feat: Add settings component
This commit adds the settings component to the frontend application. The component consists of an HTML template, a TypeScript file, and a spec file for testing. The HTML template contains a simple paragraph element with the text "settings works!". The component is standalone and does not have any imports. The component is created using Angular's @Component decorator and has a selector of 'app-settings'. The template and styles are referenced using templateUrl and styleUrl properties respectively. This addition is necessary to implement the desired settings functionality in the application.