feat: Add profile component to settings
This commit adds the profile component to the settings section of the frontend application. The profile component consists of an HTML template, a TypeScript file, and a spec file for testing. The HTML template contains a simple paragraph element with the text "profile works!". The component is standalone and does not have any imports. The component is created using Angular's @Component decorator and has a selector of 'app-profile'. The template and styles are referenced using templateUrl and styleUrl properties respectively. This addition is necessary to implement the desired profile functionality in the settings section.