feat: Add getLoggedInUser method to AuthService
This commit adds the getLoggedInUser method to the AuthService in order to retrieve the logged-in user's information. The method retrieves the user ID from the storage service and then calls the userService to get the user details. This addition is necessary to populate the profile form with the user's information in the profile settings component.