refactor: Update NavBarComponent to retrieve logged-in user from storage service
This commit updates the NavBarComponent to retrieve the logged-in user from the storage service instead of initializing it as null. The loggedInUser property is now assigned the value returned by the getUser() method of the storage service. This change ensures that the logged-in user is correctly displayed in the navigation bar component.