refactor: Update Users entity to handle missing first or last name in setFullName()
This commit updates the Users entity to handle cases where the first or last name is missing in the setFullName() method. If either the firstName or lastName is missing, the fullName is set to the available name. This change ensures that the fullName property is correctly populated even if one of the names is missing.