feat: Add follow user API endpoint
This commit adds the follow user API endpoint to the follows module. It includes the creation of the `follow user.bru` file, which defines the HTTP request for following a user. The endpoint is accessible at `{{base_url}}/follows/2/follow` and requires a bearer token for authentication. This addition allows users to follow other users and perform actions related to following.