feat: Add unfollow user API endpoint
The code changes add the unfollow user API endpoint to the follows module. It includes the creation of the `unfollow user.bru` file, which defines the HTTP request for unfollowing a user. The endpoint is accessible at `{{base_url}}/follows/2/unfollow` and requires a bearer token for authentication. This addition allows users to unfollow other users and perform actions related to unfollowing.