refactor: Add endpoint to check if the logged-in user is following another user
This commit adds a new endpoint to the follows service: '/user/:userId/following/:otherUserId'. This endpoint allows the logged-in user to check if they are following another user. By providing this functionality, we enhance the follows feature and improve the user experience by allowing users to easily determine their relationship with other users.