refactor: Update package.json to include sqlite3 dependency
This commit adds the "sqlite3" dependency to the "package.json" file. This change is necessary to use SQLite as the database for the application. By including the "sqlite3" dependency, the application will have the required package to interact with SQLite.

Code changes:
- Added "sqlite3" dependency to "package.json"

This commit improves the application's database configuration by including the necessary dependency for SQLite.