Refactor LikesService to use transaction for creating and deleting likes
Update the LikesService to use a transaction when creating and deleting likes. This ensures that the operations are atomic and consistent, preventing any data inconsistencies. By using a transaction, we can rollback the changes if an error occurs during the process. This improves the reliability and integrity of the like functionality.