refactor: Add dummy prefix to user email in user.factory.ts
Add a dummy prefix to the user email in the user.factory.ts file. This allows for easy identification and deletion of dummy users in the database. The dummy prefix is added using the format `dummy.{email}`. To find all dummy users, use the query: `select * from users where email like 'dummy.%'`.