refactor: Update seeder to populate posts with user-specific data
This commit refactors the `PopulizeTables1716408463154` seeder to populate the `posts` table with user-specific data. It updates the `seedPostsTable` method to create and save posts for each user in the `users` array. Each user will have 2 posts created. This improves the realism of the dummy data and provides more meaningful data for testing and development purposes.