feat: Add @nestjs/platform-socket.io and @nestjs/websockets dependencies
This commit adds the @nestjs/platform-socket.io and @nestjs/websockets dependencies to the backend/package.json file. These dependencies are required for implementing real-time communication using websockets in the application. By including these dependencies, the application will be able to utilize the socket.io library and leverage websockets for bidirectional communication between the server and clients.