feat: Add CORS_ORIGINS constant for allowed origins
This commit adds the CORS_ORIGINS constant to the common/constants/index.ts file. The constant includes a list of allowed origins for Cross-Origin Resource Sharing (CORS) configuration. It is used in multiple places throughout the application to specify the origins that are allowed to make requests to the server. By centralizing the list of allowed origins in a constant, it improves maintainability and makes it easier to update the list in the future.