feat: Add MessageEntity to entities and migrations
This commit adds the MessageEntity to the entities and migrations configuration files. It includes the import statement for the MessageEntity in the entities.ts file and adds the migration file for creating the messages table in the migrations.ts file. This change ensures that the MessageEntity is included in the database schema and can be used for storing and retrieving messages in the application.