chore: Update base URL for production environment
This commit updates the base URL in the frontend/src/environments/environment.production.ts file to point to the new backend server for the production environment. The previous URL has been replaced with the correct URL for the social media production environment. This change ensures that the frontend application connects to the correct backend API endpoints in the production environment.